+++ src/mlpro/gt/__init__.py
@@ -1,1 +1,2 @@
 from mlpro.gt.models import *
+from mlpro.gt.dp import *
